+++ dtor/tests/target-test/dtor/wasm32-unknown-unknown.rs
@@ -10,10 +10,11 @@ fn foo() {
             static __CTOR_PRIVATE_REF: unsafe extern "C" fn() =
                 {
                     unsafe extern "C" fn __ctor_private() {
-                        #[cfg(all(target_family = "wasm", target_os = "unknown"))]
                         {
-                            static DISARMED: ::core::sync::atomic::AtomicBool = ::core::sync::atomic::AtomicBool::new(false);
-                            if DISARMED.swap(true, ::core::sync::atomic::Ordering::Relaxed) {
+                            static DISARMED: ::core::sync::atomic::AtomicBool =
+                                ::core::sync::atomic::AtomicBool::new(false);
+                            if DISARMED.swap(true,
+                                    ::core::sync::atomic::Ordering::Relaxed) {
                                 return;
                             }
                         }