+++ link-section/src/platform/windows.rs
@@ -7,7 +7,7 @@
 #[macro_export]
 macro_rules! __get_section_windows {
     // Movable sections have an additional backref section.
-    (movable, name=$ident:ident, type=$generic_ty:ty $(, aux=$aux:ident )?) => {
+    (movable, name=$name:tt, type=$generic_ty:ty) => {
         {
             use $crate::__support::Alignment;
             use $crate::__support::PtrBounds;
@@ -23,7 +23,7 @@ macro_rules! __get_section_windows {
                 )
             } else {
                 add_section_link_attribute!(
-                    item data start $ident $($aux)?
+                    item data start $name
                     #[link_section = __]
                     static __START: SyncUnsafeCell<Alignment<$generic_ty>> = SyncUnsafeCell::new(Alignment::new());
                 );
@@ -32,14 +32,14 @@ macro_rules! __get_section_windows {
                     start.cast::<u8>().add(mem::size_of::<Alignment<$generic_ty>>()) as *const()
                 };
                 add_section_link_attribute!(
-                    item data end $ident $($aux)?
+                    item data end $name
                     #[link_section = __]
                     static __END: SyncUnsafeCell<Alignment<$generic_ty>> = SyncUnsafeCell::new(Alignment::new());
                 );
                 let end = unsafe { &raw const __END as *const () };
 
                 add_section_link_attribute!(
-                    backref data start $ident $($aux)?
+                    backref data start $name
                     #[link_section = __]
                     static __REF_START: SyncUnsafeCell<Alignment<$crate::MovableBackref<$generic_ty>>> =
                         SyncUnsafeCell::new(Alignment::new());
@@ -49,7 +49,7 @@ macro_rules! __get_section_windows {
                     start.cast::<u8>().add(mem::size_of::<Alignment<$crate::MovableBackref<$generic_ty>>>()) as *const()
                 };
                 add_section_link_attribute!(
-                    backref data end $ident $($aux)?
+                    backref data end $name
                     #[link_section = __]
                     static __REF_END: SyncUnsafeCell<Alignment<$crate::MovableBackref<$generic_ty>>> =
                         SyncUnsafeCell::new(Alignment::new());
@@ -64,7 +64,7 @@ macro_rules! __get_section_windows {
         }
     };
     // Mutable sections must use UnsafeCell to match items
-    (mutable, name=$ident:ident, type=$generic_ty:ty $(, aux=$aux:ident )?) => {
+    (mutable, name=$ident:tt, type=$generic_ty:ty) => {
         {
             use $crate::__support::Alignment;
             use $crate::__support::PtrBounds;
@@ -77,7 +77,7 @@ macro_rules! __get_section_windows {
                 PtrBounds::new(::core::ptr::null(), ::core::ptr::null())
             } else {
                 add_section_link_attribute!(
-                    item data start $ident $($aux)?
+                    item data start $ident
                     #[link_section = __]
                     static __START: SyncUnsafeCell<Alignment<$generic_ty>> = SyncUnsafeCell::new(Alignment::new());
                 );
@@ -86,7 +86,7 @@ macro_rules! __get_section_windows {
                     start.cast::<u8>().add(mem::size_of::<Alignment<$generic_ty>>()) as *const()
                 };
                 add_section_link_attribute!(
-                    item data end $ident $($aux)?
+                    item data end $ident
                     #[link_section = __]
                     static __END: SyncUnsafeCell<Alignment<$generic_ty>> = SyncUnsafeCell::new(Alignment::new());
                 );
@@ -96,7 +96,7 @@ macro_rules! __get_section_windows {
             }
         }
     };
-    ($section_type:ident, name=$ident:ident, type=$generic_ty:ty $(, aux=$aux:ident )?) => {
+    ($section_type:ident, name=$ident:tt, type=$generic_ty:ty) => {
         {
             use $crate::__support::Alignment;
             use $crate::__support::PtrBounds;
@@ -108,7 +108,7 @@ macro_rules! __get_section_windows {
                 PtrBounds::new(::core::ptr::null(), ::core::ptr::null())
             } else {
                 add_section_link_attribute!(
-                    item data start $ident $($aux)?
+                    item data start $ident
                     #[link_section = __]
                     static __START: Alignment<$generic_ty> = Alignment::new();
                 );
@@ -117,7 +117,7 @@ macro_rules! __get_section_windows {
                     start.cast::<u8>().add(mem::size_of::<Alignment<$generic_ty>>()) as *const()
                 };
                 add_section_link_attribute!(
-                    item data end $ident $($aux)?
+                    item data end $ident
                     #[link_section = __]
                     static __END: Alignment<$generic_ty> = Alignment::new();
                 );
@@ -134,14 +134,14 @@ pub use crate::__get_section_windows as get_section;
 crate::__def_section_name! {
     __section_name_windows,
     {
-        data bare =>    (".data", "$") __ ();
-        data section => (".data", "$") __ ("$b");
-        data start =>   (".data", "$") __ ("$a");
-        data end =>     (".data", "$") __ ("$c");
-        code bare =>    (".text", "$") __ ();
-        code section => (".text", "$") __ ("$b");
-        code start =>   (".text", "$") __ ("$a");
-        code end =>     (".text", "$") __ ("$c");
+        data bare =>    (".data" "$") __ ();
+        data section => (".data" "$") __ ("$b");
+        data start =>   (".data" "$") __ ("$a");
+        data end =>     (".data" "$") __ ("$c");
+        code bare =>    (".text" "$") __ ();
+        code section => (".text" "$") __ ("$b");
+        code start =>   (".text" "$") __ ("$a");
+        code end =>     (".text" "$") __ ("$c");
     }
     AUXILIARY = "$d$";
     REFS = "$r$";