+++ tests/link_section/mod.rs
@@ -11,12 +11,12 @@ defer {
 $ cargo build --quiet
 *
 $ cargo run --quiet
-! LINK_SECTION: Section { name: "%{DATA}LINK_SECTION%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, byte_len: %{INT} }
+! LINK_SECTION: Section { name: "%{DATA}LINK_SEC%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, byte_len: %{INT} }
 ! link_section_function
-! TYPED_LINK_SECTION: TypedSection { name: "%{DATA}TYPED_LINK%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 2, stride: 4 }
+! TYPED_LINK_SECTION: TypedSection { name: "%{DATA}TYPED_LI%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 2, stride: 4 }
 ! address of TYPED_LINK_SECTION[0]: %{BASE16NUM}
 ! address of TYPED_LINK_SECTION[1]: %{BASE16NUM}
-! AUX_LINK_SECTION: TypedSection { name: "%{DATA}AUX_LINK_S%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 1, stride: 4 }
+! AUX_LINK_SECTION: TypedSection { name: "%{DATA}TYPED_LI%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 1, stride: 4 }
 ! aux: 1234
 ! CODE_SECTION: TypedSection { name: "%{DATA}FN_ARRAY%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 3, stride: 8 }
 ! [%{BASE16NUM}, %{BASE16NUM}, %{BASE16NUM}]
@@ -44,7 +44,7 @@ defer {
     $ cargo clean --quiet
 }
 $ cargo run --quiet
-! INTERIOR_MUT_LINK_SECTION: TypedSection { name: "%{DATA}INTERIOR_M%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 2, stride: 8 }
+! INTERIOR_MUT_LINK_SECTION: TypedSection { name: "%{DATA}INTERIOR%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 2, stride: 8 }
 unordered {
     ! item before: InteriorMutItem { value: 1, atomic: 1 }
     ! item after: InteriorMutItem { value: 1, atomic: 2 }
@@ -63,7 +63,7 @@ defer {
     $ cargo clean --quiet
 }
 $ cargo run --quiet
-! MUT_LINK_SECTION: TypedMutableSection { name: "%{DATA}MUT_LINK_S%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 5, stride: 4 }
+! MUT_LINK_SECTION: TypedMutableSection { name: "%{DATA}MUT_LINK%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 5, stride: 4 }
 """
 item: 1
 item: 2
@@ -71,13 +71,13 @@ item: 3
 item: 4
 item: 5
 """
-! AUX_MUT_LINK_SECTION: TypedMutableSection { name: "%{DATA}AUX_MUT_LI%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 3, stride: 4 }
+! AUX_MUT_LINK_SECTION: TypedMutableSection { name: "%{DATA}MUT_LINK%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 3, stride: 4 }
 """
 aux item: 1234
 aux item: 2341
 aux item: 4321
 """
-! MOVABLE_LINK_SECTION: TypedMovableSection { name: "%{DATA}MOVABLE_LI%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 4, stride: 4 }
+! MOVABLE_LINK_SECTION: TypedMovableSection { name: "%{DATA}MOVABLE%{DATA}", start: %{BASE16NUM}, end: %{BASE16NUM}, len: 4, stride: 4 }
 ! MOVABLE_BACKREFS: 4
 """
 movable item: 10
@@ -102,6 +102,7 @@ defer {
 }
 $ cargo run --quiet
 ! link-section-no-default-features:in-section
+! link-section-no-default-features:in-section-aux
 ! link-section-no-default-features:main
 "#
 );